MAINT: Split test file in two to find hanging test.
It looks like some tests in test_qt_dims do hang sometime in CI.
Though it's not clear which one with Pytest output.

This split the test into two file approximatively where the hang
happens, so we can narrow down the problem.

This may help with #5443 maybe ?
